supply_chain/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
hdfs==2.7.3
idna==3.10
kafka-python==2.0.2
PyJWT==2.10.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
urllib3==2.2.3
